# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SenAI Agentic CRM Intelligence API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/drafts/{id}": {
      "patch": {
        "tags": [
          "drafts"
        ],
        "summary": "Update Draft",
        "operationId": "update_draft_drafts__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}/approve": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Approve Draft",
        "operationId": "approve_draft_drafts__id__approve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/sentiment-trend": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Sentiment Trend Endpoint",
        "operationId": "get_sentiment_trend_endpoint_analytics_sentiment_trend_get",
        "parameters": [
          {
            "name": "sender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sender email filter",
              "title": "Sender"
            },
            "description": "Sender email filter"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Number of days to check",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days to check"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/category-breakdown": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Category Breakdown",
        "operationId": "get_category_breakdown_analytics_category_breakdown_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start range datetime",
              "title": "Start Date"
            },
            "description": "Start range datetime"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End range datetime",
              "title": "End Date"
            },
            "description": "End range datetime"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/contacts/{email}": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Get Contact Profile Endpoint",
        "operationId": "get_contact_profile_endpoint_contacts__email__get",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{email}/status": {
      "patch": {
        "tags": [
          "contacts"
        ],
        "summary": "Update Contact Status",
        "operationId": "update_contact_status_contacts__email__status_patch",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactStatusPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContactStatusPayload": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "VIP",
              "Blocked",
              "Active",
              "Churned"
            ],
            "title": "Status",
            "description": "The contact's CRM status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContactStatusPayload"
      },
      "DraftUpdatePayload": {
        "properties": {
          "proposed_content": {
            "type": "string",
            "minLength": 1,
            "title": "Proposed Content",
            "description": "The updated text content of the draft reply"
          }
        },
        "type": "object",
        "required": [
          "proposed_content"
        ],
        "title": "DraftUpdatePayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}